fix: minor bugs

+++ src/colors/colors.service.ts
@@ -1,15 +1,28 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateColorDto } from './dto/create-color.dto';
 import { ColorsRepository } from 'src/repositories/colors';
 import { ColorsFilter } from './dto/query.dto';
+import { ProductOptionsRepository } from 'src/repositories/product_options';
+import { ColorType } from 'src/common/types/enums';
 
 @Injectable()
 export class ColorsService {
-  constructor(private readonly colorsRepository: ColorsRepository) {}
+  constructor(
+    private readonly colorsRepository: ColorsRepository,
+    private readonly productsOptionsRepository: ProductOptionsRepository,
+  ) {}
   async create(createColorDto: CreateColorDto) {
     const hasCode = await this.colorsRepository.findBy({
       code: createColorDto.code,
     });
+    createColorDto.name_uz = createColorDto.name_uz
+      .trim()
+      .replace(/\s+/g, ' ')
+      .replace(/['‘’`]/g, "'");
     if (hasCode) {
       throw new ConflictException('Color already exists');
     }
@@ -21,7 +34,7 @@ export class ColorsService {
   }
 
   findOne(id: number) {
-    return `This action returns a #${id} color`;
+    return this.colorsRepository.findBy({ id });
   }
 
   async update(id: number, updateColorDto: CreateColorDto) {
@@ -32,10 +45,37 @@ export class ColorsService {
     if (hasCode && hasCode.id !== id) {
       throw new ConflictException('Color already exists');
     }
+
+    updateColorDto.name_uz = updateColorDto.name_uz
+      .trim()
+      .replace(/\s+/g, ' ')
+      .replace(/['‘’`]/g, "'");
     return this.colorsRepository.update({ id }, updateColorDto);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const hasColor = await this.colorsRepository.findBy({
+      id,
+    });
+    if (!hasColor) {
+      throw new NotFoundException('Color not found');
+    }
+
+    let query = {};
+    if (hasColor.type == ColorType.DECOR) {
+      query = {
+        decor_color_id: id,
+      };
+    } else {
+      query = {
+        main_color_id: id,
+      };
+    }
+    const hasProduct = await this.productsOptionsRepository.findBy(query);
+
+    if (hasProduct) {
+      throw new ConflictException('Color has products');
+    }
     return this.colorsRepository.delete({ id });
   }
 }
