+++ src/measurement/measurement.service.ts
@@ -1,21 +1,33 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  UnprocessableEntityException,
+} from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { MeasurementsFilter } from './dto/query.dto';
 import { MeasurementsRepository } from 'src/repositories/measurement';
+import { ProductsRepository } from 'src/repositories/products';
+import { ProductOptionsRepository } from 'src/repositories/product_options';
 
 @Injectable()
 export class MeasurementsService {
   constructor(
     private readonly measurementsRepository: MeasurementsRepository,
+    private readonly productsOptionsRepository: ProductOptionsRepository,
   ) {}
 
   async create(payload: CreateCategoryDto) {
+    payload.name_uz = payload.name_uz
+      .trim()
+      .replace(/\s+/g, ' ')
+      .replace(/['‘’`]/g, "'");
+
     const hasCategory = await this.measurementsRepository.findBy({
-      name: payload.name_uz,
+      name_uz: payload.name_uz,
     });
 
     if (hasCategory) {
-      throw new Error('Category already exists');
+      throw new ConflictException('Category already exists');
     }
 
     return this.measurementsRepository.create(payload);
@@ -30,17 +42,29 @@ export class MeasurementsService {
   }
 
   async update(id: number, payload: CreateCategoryDto) {
+    payload.name_uz = payload.name_uz
+      .trim()
+      .replace(/\s+/g, ' ')
+      .replace(/['‘’`]/g, "'");
+
     const hasDuplicate = await this.measurementsRepository.findBy({
-      name: payload.name_uz,
+      name_uz: payload.name_uz,
     });
 
     if (hasDuplicate && hasDuplicate.id !== id) {
-      throw new Error('Category already exists');
+      throw new ConflictException('Category already exists');
     }
     return this.measurementsRepository.update({ id }, payload);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const hasProducts = await this.productsOptionsRepository.findBy({
+      measurement_id: id,
+    });
+
+    if (hasProducts) {
+      throw new UnprocessableEntityException('Category has products');
+    }
     return this.measurementsRepository.delete({ id });
   }
 }
